Fix combining lines across files
Previously the last line of a file could become concatenated with
the first line of the next, if it did not end with a newline char.
This restricts the search to lines with whole words, which as well
as reducing the overall search area (by around half in a test repo),
is more representative of realistic searches (for words).