Try switching 'Ag' command to files + contents
This is to support a more grep-like behaviour, where it's possible
to choosr the files and content to search in the same command. For
example, it supports searches:

- For contents containing 'download' *and*
- In files containing 'app/'

While there is a risk of some false positives (in this case, if the
contents also contained '/app'), the aim of trying this is to see if
it's more useful than not.