Revert "Fix combining lines across files"
This reverts commit 825589fa2025aeed2c49146b910eb39e4d18a62e.

It turns out the '-w' flag applies to the search pattern used for
ag, which in this case is '.'. As a consequence, only certain lines
were returned, which contained a single character word. There does
not seem to be a way to fix both issues, but reverting seems best
to try and evaluate how bad the original issue was.