Add fugitive/rhubarb plugin for Git browsing
This makes it easy to browse to a file in GitHub in order to:

- Explore its history further.

- Generate a stable link to use as a reference in a commit.

Visually selecting one or more lines causes the URL to be for the
last commit that affected them.

"ruanyl/vim-gh-line" does something similar, but only opens URLs
for the latest commit overall, which is unlikely to be stable in
the context of generating references as part of a new commit on a
branch - the commits may change.