Support listing relevant files to change (grep/rg)
This makes it easier to have an entire change process within Vim,
rather than having to go back to a list of terminal results to see
which files need to be changed, often in some complex way beyond a
simple find/replace. Indeed, the quickfix list makes it easy to go
directly to each file, just by pressing "Enter" on it.