Fix and tweak background for line / selection
This was broken as it came before the command to set colorscheme,
which set the background colour to some shade of grey.

Normally highlighting the current line is a peripheral concern -
grey was sufficient due to most lines having enough surrounding
black for the grey to be apparent - but the highlighting is hard
to see in quickfix lists as there is less empty space.

Quickfix actually has two navigation methods:

- Using normal up/down keys. This behaves the same visually as in
a normal text file (grey).

- Using quickfix navigation keys e.g. ":cn". This highlights the
current line clearly (yellow).

Since the quickfix keys are less convenient, this changes both of
the methods to highlight in the same way: color code 17 is a dark
shade of blue that seems to be good compromise by being visually
distinct but not to obtrusive (it's only a highlight).

For consistency, this also modifies the visual selection - as an
extension of hovering over a single line.