Synchronise colours for fzf Ag and Files
For Ag it was previously impossible to see the matched text versus
the other text for the selected item; this fixes that by switching
to the "green on purple" style of the previous, homegrown functions.

This also removes "--exact" on finding files as this is more useful
versus the potential of having lower quality results appear.