Try switching leader key to space
The width of the space bar allows for two handed-operation and in
this way makes more <leader><key> combinations practical.

The BufExplorer mapping is updated to "," as "." is now a bit far
to reach. I've also disabled the automatic mapping of <space> in
Tagbar so that the toggle mapping continues to work.