Try namespacing lesser used shortcuts with ","
Each namespace is based on the "primary" key map associated with
a particular plugin e.g. "f" for fzf.

Making the "primary" key the namespace would cause a delay in the
use of the primary shortcut due to ambiguity, so prefixing a ","
is necessary to get the best of both. Having to type "," as well
as the namespace key shouldn't be an issue for these shortcuts.