Disable swap files to avoid multi-edit check
This is something that happens fairly often and is really designed
for multi-user systems - Vim will still warn if a reopened file is
being edited when changes are saved.

The other aspect of swap files is recovery, but I don't recall ever
using this in practice - the combination of version control and just
saving files frequently should be enough.

Note that "directory" is specifically the location for swap files -
backup and undo ("~") files go elsewhere.