Fix GitFiles not working from missing files
Steps to reproduce:

- Commit a new file in a directory.
- Open the file as the current buffer.
- Checkout a previous commit.
- Try to open the file finder.

This would show an error of "Not a git repo." due to the gitfiles
function trying to work out the root directory based on the current
file [^1]. Using a manual "git ls-files" command fixes the issue.

At the same time, it makes sense to re-add support for finding files
outside of Git repos, as this can occasionally be useful.

[^1]: https://github.com/junegunn/fzf.vim/blob/master/autoload/fzf/vim.vim#L618