Increase highlight for selection colours
For quickfix lists this makes it possible to distinguish between the
active line and the selected one.

For visual selection it's just easier to see what's actually being
selected. Unlike the curosr line, where much of the highlight is over
black empty space, visual selection only highlights the actual text,
which contrasts much less with the highlight colour.