Make Fzf Rg command argunents editable later on
Previously any arguments passed to :Rg would be passed on to the
"rg" command itself. Although this is more efficient, in practice
it makes more sense to pass the query to Fzf:

- The initial filter is visible as feedback for what was entered.
This is useful if Rg is being invoked from a shortcut.

- The initial filter can be refined by editing. This is useful if
Rg is being called with speculative search terms.

While there were no active usecases for the old behaviour, the new
behaviour has several shortcuts to utilise it:

- Pre-populate with "def | class <cword>(" - find definitions

- Pre-populate with ".<cword>(" - find usages