Enable quickfix listing of Fzf file results
This also requires the "--multi" option to support selection of all
files. Like "with_preview", "fzf#wrap" extends the passed options;
it adds a default "edit path" sink as well as bindings for keys in
"fzf_actions" [^1] i.e. the quickfix feature.

I did investigate if it's possible to refactor the "Rg" function to
have a similar structure, but the default "edit path" sink doesn't
support line numbers (it thinks they're part of the path) [^2]. If
more functions were exposed in fzf#vim then it might be possible to
specify a custom sink, but it's not simple [^3].

[^1]: https://github.com/junegunn/fzf/blob/80b88463188ff78719f23cb2e1dca5f27227879e/plugin/fzf.vim#L428-L435
[^2]: https://github.com/junegunn/fzf/blob/80b88463188ff78719f23cb2e1dca5f27227879e/plugin/fzf.vim#L328
[^3]: https://github.com/junegunn/fzf.vim/blob/master/autoload/fzf/vim.vim#L190