Specify colour scheme for Coc suggestions
With the original scheme it was hard to discern the selected item
(very dark grey) as well as the overall list of suggestions (black)
from the file text. Also, the highlight of matching search text (dark
blue) was less prominent than unmatched text (light blue).

Using purples and blues is consistent with Fzf settings [^1].

[^1]: https://github.com/benthorner/vim-rc/blob/90df305/vimrc#L79-L80